microservicios/notification_service/requirements.txt
flask
requests
pulsar-client
